Small commenting changes
- Added small comments littered throughout.
- Removed redundant messageHandlers.js in Content Script.
